# api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: validation error
        type: string
    type: object
...
  handlers.UpdateProfilePhotoRequest:
    properties:
      fileId:
        type: string
    type: object
...
  presenter.PaginatedUsersResponse:
    properties:
      items:
        items:
          $ref: "#/definitions/presenter.UserResponse"
        type: array
      meta:
        $ref: "#/definitions/presenter.PaginationMetaResponse"
    type: object
  presenter.PaginationMetaResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      search:
        type: string
      sortBy:
        type: string
      sortOrder:
        type: string
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  presenter.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      mustChangePassword:
        type: boolean
      name:
        type: string
      preferredLocale:
        allOf:
          - $ref: "#/definitions/user.Locale"
        enum:
          - fr
          - en
        example: fr
      profilePhotoFileId:
        type: string
      profilePhotoUrl:
        type: string
      role:
        allOf:
          - $ref: "#/definitions/user.Role"
        enum:
          - admin
          - user
        example: user
      updatedAt:
        type: string
    type: object
  user.Locale:
    enum:
      - fr
      - en
    type: string
    x-enum-varnames:
      - LocaleFR
      - LocaleEN
  user.Role:
    enum:
      - admin
      - user
    type: string
    x-enum-varnames:
      - RoleAdmin
      - RoleUser
info:
  contact: {}
  description:
    API server built with Gin, Gorm, Postgres, JWT auth, refresh tokens,
    and Swaggo.
  title: Boilerplate Go Vue API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: Returns all users. Admin only.
      parameters:
        - description: Page number
          in: query
          name: page
          type: integer
        - description: Page size
          in: query
          name: limit
          type: integer
        - description: Search on name, email, role
          in: query
          name: search
          type: string
        - description: Sort field
          enum:
            - createdAt
            - updatedAt
            - name
            - email
            - role
          in: query
          name: sortBy
          type: string
        - description: Sort order
          enum:
            - asc
            - desc
          in: query
          name: sortOrder
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/presenter.PaginatedUsersResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List users
      tags:
        - users
  /api/v1/users/{id}:
    get:
      description: Returns a single user by id. Admin only.
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/presenter.UserResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Get user by ID
      tags:
        - users
  /api/v1/users/me:
    get:
      description: Returns the authenticated user profile.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/presenter.UserResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Current user
      tags:
        - users
  /api/v1/users/me/profile-photo:
    patch:
      consumes:
        - application/json
      description:
        Assigns or clears the authenticated user's profile photo using
        an uploaded file id.
      parameters:
        - description: Profile photo payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.UpdateProfilePhotoRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/presenter.UserResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Update current user profile photo
      tags:
        - users
securityDefinitions:
  BearerAuth:
    description: Use `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"